backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
requests
python-multipart
openpyxl
